requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2018.1.18
chardet==2.3.0
Django==1.11
django-admin==1.1.1
django-admin-tools==0.8.1
django-allauth==0.34.0
django-cors-headers==2.2.0
django-excel-response2==2.0.8
django-filter==1.1.0
django-jet==1.0.7
django-material==1.1.0
django-rest-auth==0.9.3
django-six==1.0.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docx2txt==0.6
drf-generators==0.3.0
EbookLib==0.15
enum==0.4.6
et-xmlfile==1.0.1
html2pdf==0.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
jdcal==1.3
Jinja2==2.11.dev0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
MySQL-python==1.2.5
numpy==1.14.4
oauthlib==2.0.6
openpyxl==2.5.0
pandas==0.23.0
pdfminer==20140328
Pillow==5.0.0
pocketsphinx==0.1.3
Postdown==1.0.11
PyJWT==1.5.3
PyMySQL==0.8.1
PyPDF2==1.26.0
python-dateutil==2.7.3
python-openid==2.2.5
python-pptx==0.6.5
pytz==2017.3
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
screen==1.0.1
six==1.10.0
SpeechRecognition==3.6.3
textract==1.6.1
urllib3==1.22
virtualenv==15.1.0
webencodings==0.5.1
xhtml2pdf==0.2.1
xlrd==1.0.0
XlsxWriter==1.0.5
xlwt==1.3.0
xmltodict==0.11.0
